% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{calWeights}
\alias{calWeights}
\title{Calculate weight matrix}
\usage{
calWeights(SpaCET_obj, radius = 200, k = NULL, sigma = 100, diagAsZero = TRUE)
}
\arguments{
\item{SpaCET_obj}{A SpaCET object.}

\item{radius}{Radius cut off (unit: um).}

\item{k}{The maximum number of nearest neighbours to compute. The default value is all within radius.}

\item{sigma}{Free parameter for RBF kernel.}

\item{diagAsZero}{Indicate whether to mask the effect in the same spot.}
}
\value{
A SpaCET object
}
\description{
Calculate weight matrix based on the Radial Basis Function (RBF) kernel.
}
\examples{
W <- calWeights(SpaCET_obj)

}
