% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert.Seurat}
\alias{convert.Seurat}
\title{Convert Seurat to SpaCET}
\usage{
convert.Seurat(Seurat_obj, platform, visiumPath = NULL, organism = "human")
}
\arguments{
\item{Seurat_obj}{A Seurat object.}

\item{platform}{A character string indicating the platform, i.e., "Visium", "OldST", or "Slide-Seq". "OldST" is the early in situ capturing method from which "Visium" was developed.}

\item{visiumPath}{Path to the Space Ranger output folder (Optional). If setting, this function will retrieve more information from the raw output.}

\item{organism}{Organism of the sample, e.g., human or mouse.}
}
\value{
A SpaCET object.
}
\description{
Convert a Seurat object to a SpaCET object.
}
\examples{
visiumPath <- file.path(system.file(package = "SpaCET"), "extdata/Visium_BC")
Seurat_obj <- Seurat::Load10X_Spatial(data.dir = visiumPath)
SpaCET_obj <- convert.Seurat(Seurat_obj, platform = "Visium")
SpaCET_obj <- SpaCET.quality.control(SpaCET_obj)
SpaCET.visualize.spatialFeature(SpaCET_obj, spatialType = "QualityControl", spatialFeatures = c("UMI","Gene"))

}
