% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create.SpaCE.object}
\alias{create.SpaCE.object}
\title{Create an SpaCE object}
\usage{
create.SpaCE.object(counts, spotCoordinates, imagePath, platform)
}
\arguments{
\item{counts}{Count matrix with gene name (row) x spot ID (column)}

\item{spotCoordinates}{Spot coordinate matrix with spot ID (column) x coordinates (row). This matrix should include two columns, i,e., X and Y coordinates, respectively, which represent the position of spots in H&E image.}

\item{imagePath}{Path to the H&E image file. Can be NA if it is not available.}

\item{platform}{A character string indicating the platform, i.e., "Visium", "oldST", or "Slide-Seq".}
}
\value{
An SpaCE object
}
\description{
Read an ST dataset to create an SpaCE object.
}
\details{
To create an SpaCE object, user need to input four parameters, i.e., "counts", "spotCoordinates", "imageFile", and "platform".
However, if analyzing Visium data, please use create.SpaCE.object.10X to read data.
}
\examples{
SpaCE_obj <- create.SpaCE.object(counts,spotCoordinates,imagePath,platform)

}
