% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.SpaCE.R
\name{create.SpaCE}
\alias{create.SpaCE}
\title{FUNCTION_TITLE}
\usage{
create.SpaCE(seqPath, imagePath)
}
\arguments{
\item{seqPath}{PARAM_DESCRIPTION}

\item{imagePath}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[Matrix]{externalFormats}}
\code{\link[jsonlite]{toJSON, fromJSON}}
}
