% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create.SpaCET.object.10X}
\alias{create.SpaCET.object.10X}
\title{Create a SpaCET object from 10X Genomics platforms}
\usage{
create.SpaCET.object.10X(
  dataPath = NULL,
  visiumPath = NULL,
  platform = "Visium",
  organism = "human"
)
}
\arguments{
\item{dataPath}{Path to a Visium, Visium HD, or Xenium output directory.
See \emph{Details} for the directory expected for each platform.}

\item{visiumPath}{Path to the Space Ranger output folder. This is a deprecated parameter. Please use `dataPath` instead.}

\item{platform}{A character string indicating the platform, i.e., "Visium", "VisiumHD", or "Xenium".}

\item{organism}{Organism of the sample, e.g., human or mouse.}
}
\value{
A SpaCET object.
}
\description{
Read an ST dataset to create a SpaCET object.
}
\details{
The directory supplied through \code{dataPath} depends on the platform:
\itemize{
  \item{\strong{Visium}: the Space Ranger output directory.}
  \item{\strong{Binned Visium HD}: one
  \code{binned_outputs/square_*um} directory.}
  \item{\strong{Segmented Visium HD}: the
  \code{segmented_outputs} directory.}
  \item{\strong{Xenium}: the Xenium output directory.}
}
}
\examples{
visiumPath <- file.path(system.file(package = "SpaCET"), "extdata/Visium_BC")
SpaCET_obj <- create.SpaCET.object.10X(dataPath = visiumPath, platform = "Visium", organism="human")

}
