% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create.SpaCET.object.CosMx}
\alias{create.SpaCET.object.CosMx}
\title{Create a SpaCET object from NanoString CosMx}
\usage{
create.SpaCET.object.CosMx(cosmxPath, fov = NULL, organism = "human")
}
\arguments{
\item{cosmxPath}{Path to the CosMx output folder containing the expression matrix and metadata.}

\item{fov}{Field of view indices to include. Default NULL loads all FOVs.}

\item{organism}{Species, either "human" (default) or "mouse".}
}
\value{
A SpaCET object.
}
\description{
Read NanoString CosMx SMI output into a SpaCET object.
}
\examples{
\dontrun{
SpaCET_obj <- create.SpaCET.object.CosMx(cosmxPath = "/path/to/cosmx_output")
}

}
