% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create.SpaCET.object}
\alias{create.SpaCET.object}
\title{Create a SpaCET object}
\usage{
create.SpaCET.object(
  counts,
  spotCoordinates,
  metaData = NULL,
  imagePath = NA,
  platform,
  organism = "human"
)
}
\arguments{
\item{counts}{Count matrix with gene name (row) x spot ID (column).}

\item{spotCoordinates}{Spot coordinate matrix with spot ID (row) x coordinates (column). This matrix should include two columns, i.e., X and Y coordinates, respectively, which represent the position of spots in H&E image.}

\item{metaData}{Meta data matrix with spot ID (row) x meta information (column).}

\item{imagePath}{Path to the H&E image file. Can be NA if it is not available.}

\item{platform}{A character string indicating the platform, i.e., "Visium", "OldST", or "Slide-Seq". "OldST" is the early in situ capturing method from which "Visium" was developed.}

\item{organism}{Organism of the sample, e.g., human or mouse.}
}
\value{
A SpaCET object.
}
\description{
Read an ST dataset to create a SpaCET object.
}
\details{
To create a SpaCET object, user need to input four parameters, i.e., "counts", "spotCoordinates", "imagePath", and "platform".
However, if analyzing the Visium data, it is more easy to use `create.SpaCET.object.10X` to read ST data.
}
\examples{
SpaCET_obj <- create.SpaCET.object(counts, spotCoordinates, imagePath, platform)

}
