% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create.SpaCET.object.Xenium}
\alias{create.SpaCET.object.Xenium}
\title{Create a SpaCET object from 10x Xenium}
\usage{
create.SpaCET.object.Xenium(xeniumPath, organism = "human")
}
\arguments{
\item{xeniumPath}{Path to the Xenium output folder.}

\item{organism}{Species, either "human" (default) or "mouse".}
}
\value{
A SpaCET object.
}
\description{
Read 10x Genomics Xenium output into a SpaCET object.
}
\examples{
\dontrun{
SpaCET_obj <- create.SpaCET.object.Xenium(xeniumPath = "/path/to/xenium_output")
}

}
