% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.ST.R
\name{deconvolve.ST}
\alias{deconvolve.ST}
\title{Deconvolution}
\usage{
deconvolve.ST(ST, tempFilePath)
}
\arguments{
\item{ST}{An SpaCE object}

\item{tempFilePath}{Path to a temporary folder for storing infercnv results.}
}
\value{
An SpaCE object
}
\description{
Deconvovle tumor ST data
}
\details{
This function first estimates cancer cell abundance and clonal substructures through modeling the segmental copy number variations across ST spots. A constrained regression model can calibrate local tissue densities and determine stromal and immune cell lineage hierarchies.
}
\examples{
ST <- deconvolve.ST(ST,tempFilePath)
}
