% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.ST.R
\name{inferMalignant}
\alias{inferMalignant}
\title{FUNCTION_TITLE}
\usage{
inferMalignant(ST, Ref, infercnv_dir)
}
\arguments{
\item{ST}{PARAM_DESCRIPTION}

\item{Ref}{PARAM_DESCRIPTION}

\item{infercnv_dir}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[GSVA]{gsva}},\code{\link[GSVA]{igsva}},\code{\link[GSVA]{GSVA}}
\code{\link[infercnv]{c("CreateInfercnvObject", "CreateInfercnvObject")}},\code{\link[infercnv]{run}}
}
