% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{read.gmt}
\alias{read.gmt}
\title{Read a gmt file.}
\usage{
read.gmt(gmtPath)
}
\arguments{
\item{gmtPath}{Path to a gmt file.}
}
\value{
A gene set list
}
\description{
Read a gmt file as a gene set list.
}
\examples{
gmt <- read.gmt(gmtPath)

}
