% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{class}
\name{SpaCE-class}
\alias{SpaCE-class}
\title{Define the SpaCE object}
\description{
Define the SpaCE object
}
\section{Slots}{

\describe{
\item{\code{input}}{The input data}

\item{\code{results}}{The results}
}}

