% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{SpaCE.CCI.cellTypePair}
\alias{SpaCE.CCI.cellTypePair}
\title{Ligand-Receptor analysis for a co-localized cell-type pair}
\usage{
SpaCE.CCI.cellTypePair(SpaCE_obj, cellTypePair)
}
\arguments{
\item{SpaCE_obj}{An SpaCE object.}

\item{cellTypePair}{Cancer type of this tumor ST dataset.}
}
\value{
An SpaCE object
}
\description{
Test the co-expression of ligands and receptors within the same ST spot for the co-localized cell-type pairs.
}
\examples{
SpaCE.CCI.cellTypePair(SpaCE_obj,cellTypePair=c("CAF","Macrophage M2"))
}
