% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{SpaCE.CCI.colocalization}
\alias{SpaCE.CCI.colocalization}
\title{Cell-cell colocalization analysis}
\usage{
SpaCE.CCI.colocalization(SpaCE_obj)
}
\arguments{
\item{SpaCE_obj}{An SpaCE object.}
}
\value{
An SpaCE object
}
\description{
Calculate the cell-type pair colocalization by using Spearman correlation analysis.
}
\examples{
SpaCE_obj <- SpaCE.CCI.colocalization(SpaCE_obj)
}
