% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{SpaCE.deconvolution.malignant}
\alias{SpaCE.deconvolution.malignant}
\title{Deconvolve malignant cell fraction}
\usage{
SpaCE.deconvolution.malignant(SpaCE_obj, malignantCutoff = 0.7, coreNo = 8)
}
\arguments{
\item{SpaCE_obj}{An SpaCE object.}

\item{malignantCutoff}{Fraction cutoff for defining spots with high abundant malignant cells.}

\item{coreNo}{Core number.}
}
\value{
An SpaCE object
}
\description{
Explore different cancer cell state in tumor ST dataset.
}
\examples{
SpaCE_obj <- SpaCE.deconvolution.malignant(SpaCE_obj)
}
