% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{SpaCE.visualize.colocalization}
\alias{SpaCE.visualize.colocalization}
\title{Cell-cell colocalization visualization}
\usage{
SpaCE.visualize.colocalization(SpaCE_obj)
}
\arguments{
\item{SpaCE_obj}{An SpaCE object.}
}
\value{
An ggplot object
}
\description{
Visualize the cell-type pair colocalization.
}
\examples{
SpaCE.visualize.colocalization(SpaCE_obj)
}
