% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SpaCE.visualize.LRNetworkScore}
\alias{SpaCE.visualize.LRNetworkScore}
\title{Ligand-Receptor network score visualization}
\usage{
SpaCE.visualize.LRNetworkScore(
  SpaCE_obj,
  colors = c("black", "black", "black", "blue", "blue", "blue", "blue", "cyan", "cyan",
    "yellow"),
  imageBg = TRUE
)
}
\arguments{
\item{SpaCE_obj}{An SpaCE object.}

\item{colors}{Legend color scale, Default: c("black","black","black","blue","blue","blue","blue","cyan","cyan","yellow").}
}
\value{
A ggplot2 object
}
\description{
Visualize L-R network score in ST dataset.
}
\examples{
SpaCE.visualize.LRNetworkScore(SpaCE_obj)
}
