% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SpaCE.visualize.metrics}
\alias{SpaCE.visualize.metrics}
\title{Quality control metrics visualization}
\usage{
SpaCE.visualize.metrics(
  SpaCE_obj,
  itemQC = c("UMI", "Gene"),
  colors = c("lightblue", "blue", "darkblue"),
  imageBg = TRUE
)
}
\arguments{
\item{SpaCE_obj}{An SpaCE object.}

\item{itemQC}{Item for quality control metrics. i.e., "UMI" or "gene".}

\item{colors}{Legend color scale, Default: c("blue", "yellow", "red").}

\item{imageBg}{logical: should the image be shown?}
}
\value{
A ggplot2 object
}
\description{
Visualize quality control metrics in ST dataset.
}
\examples{
SpaCE.visualize.metrics(SpaCE_obj, itemQC="UMI")
}
