% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{class}
\name{SpaCET-class}
\alias{SpaCET-class}
\title{Define the SpaCET object}
\description{
Define the SpaCET object
}
\section{Slots}{

\describe{
\item{\code{input}}{The input data}

\item{\code{results}}{The results}
}}

