% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{SpaCET.calculate.mostAbundantCellType}
\alias{SpaCET.calculate.mostAbundantCellType}
\title{Calculate the most abundant cell type}
\usage{
SpaCET.calculate.mostAbundantCellType(SpaCET_obj)
}
\arguments{
\item{SpaCET_obj}{A SpaCET object containing cell-type deconvolution results.}
}
\value{
A SpaCET object.
}
\description{
Identify the most abundant major lineage and sublineage.
}
\details{
The function compares the estimated fractions in
\code{SpaCET_obj@results$deconvolution$propMat}. Candidate cell types are
obtained from \code{SpaCET_obj@results$deconvolution$Ref$lineageTree}.
The \code{Unidentifiable} component is not treated as a cell type. If two
cell types have the same fraction, the first one in the reference lineage
tree is selected.
}
\examples{
\dontrun{
SpaCET_obj <- SpaCET.calculate.mostAbundantCellType(SpaCET_obj)
}

}
