% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{SpaCET.CCI.cellTypePair}
\alias{SpaCET.CCI.cellTypePair}
\title{Ligand-Receptor analysis for a co-localized cell-type pair}
\usage{
SpaCET.CCI.cellTypePair(SpaCET_obj, cellTypePair)
}
\arguments{
\item{SpaCET_obj}{A SpaCET object.}

\item{cellTypePair}{A pair of cell-types.}
}
\value{
A SpaCET object.
}
\description{
Test the co-expression of ligands and receptors within the same ST spot for the co-localized cell-type pairs.
}
\examples{
SpaCET_obj <- SpaCET.CCI.cellTypePair(SpaCET_obj, cellTypePair=c("CAF","Macrophage M2"))

}
