% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{SpaCET.CCI.colocalization}
\alias{SpaCET.CCI.colocalization}
\title{Cell-cell colocalization analysis}
\usage{
SpaCET.CCI.colocalization(SpaCET_obj)
}
\arguments{
\item{SpaCET_obj}{An SpaCET object.}
}
\value{
An SpaCET object
}
\description{
Calculate the cell-type pair colocalization by using Spearman correlation analysis.
}
\examples{
SpaCET_obj <- SpaCET.CCI.colocalization(SpaCET_obj)
}
