% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{SpaCET.CCI.LRNetworkScore}
\alias{SpaCET.CCI.LRNetworkScore}
\title{Ligand-Receptor interaction enrichment analysis}
\usage{
SpaCET.CCI.LRNetworkScore(SpaCET_obj, coreNo = 8)
}
\arguments{
\item{SpaCET_obj}{An SpaCET object.}

\item{coreNo}{Core number in parallel computation.}
}
\value{
An SpaCET object.
}
\description{
Calculate the overall intensity of L-R interactions within ST spots.
}
\examples{
SpaCET_obj <- SpaCET.CCI.LRNetworkScore(SpaCET_obj, coreNo=8)
SpaCET.visualize.spatialFeature(SpaCET_obj, spatialType = "LRNetworkScore", spatialFeatures=c("Network_Score","Network_Score_pv"))

}
