% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{SpaCET.combine.interface}
\alias{SpaCET.combine.interface}
\title{Combine interaction spots to interface}
\usage{
SpaCET.combine.interface(SpaCET_obj, cellTypePair)
}
\arguments{
\item{SpaCET_obj}{A SpaCET object.}

\item{cellTypePair}{A pair of cell types.}
}
\value{
A SpaCET object.
}
\description{
Demonstrate the spatial position of interaction spots at the tumor microenvironment.
}
\examples{
SpaCET_obj <- SpaCET.combine.interface(SpaCET_obj,cellTypePair=c("CAF", "Macrophage M2"))
SpaCET.visualize.spatialFeature(SpaCET_obj, spatialType = "Interface", spatialFeature = "Interaction")

}
