% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{SpaCET.distance.to.interface}
\alias{SpaCET.distance.to.interface}
\title{Distance to tumor border}
\usage{
SpaCET.distance.to.interface(
  SpaCET_obj,
  cellTypePair = c("CAF", "Macrophage M2"),
  nPermutation = 1000
)
}
\arguments{
\item{SpaCET_obj}{A SpaCET object.}

\item{cellTypePair}{A pair of cell types.}

\item{nPermutation}{Permutation number.}
}
\value{
A ggplot object.
}
\description{
Calculate the distance of cell-cell interactions to tumor-immune interface.
}
\examples{
SpaCET.distance.to.interface(SpaCET_obj, cellTypePair=c("CAF", "Macrophage M2"))

}
