% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{SpaCET.GeneSetScore}
\alias{SpaCET.GeneSetScore}
\title{Calculate gene set score for each spot}
\usage{
SpaCET.GeneSetScore(SpaCET_obj, GeneSets)
}
\arguments{
\item{SpaCET_obj}{An SpaCET object.}

\item{GeneSets}{A string for in-house gene sets, or a list object for user-defined gene sets. See details.}
}
\value{
An SpaCET object
}
\description{
Calculate spots' gene set score from the in-house or user-defined gene sets.
}
\details{
1) Set `GeneSets` as "Hallmark", "CancerCellState", or "TLS" to use the in-house gene sets.
"Hallmark": https://www.gsea-msigdb.org/gsea/msigdb/human/collections.jsp#H

"CancerCellState": https://www.nature.com/articles/s41588-022-01141-9

"TLS": https://www.researchsquare.com/article/rs-3921508

2) Set `GeneSets` as a list to use the user-defined gene sets. Each entry should be a vector of gene symbols.

The function `ScoreSignatures_UCell` from `UCell` is used to calculate gene-set scores.
https://bioconductor.org/packages/release/bioc/html/UCell.html.
}
\examples{
SpaCET_obj <- SpaCET.GeneSetScore(SpaCET_obj, GeneSets="Hallmark")

}
