% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{SpaCET.SpatialCorrelation}
\alias{SpaCET.SpatialCorrelation}
\title{Calculate spatial correlation}
\usage{
SpaCET.SpatialCorrelation(
  SpaCET_obj,
  mode,
  item = NULL,
  W,
  nPermutation = 1000
)
}
\arguments{
\item{SpaCET_obj}{A SpaCET object.}

\item{mode}{Mode of spatial correlation, i.e., "univariate", "bivariate", "pairwise".}

\item{item}{A vector or two-column matrix. See details.}

\item{W}{Weight matrix.}

\item{nPermutation}{Permutation number.}
}
\value{
A SpaCET object
}
\description{
Calculate spatial correlations in both univariate and bivariate settings.
}
\details{
1) when mode = "univariate", set `item` as NULL or a gene vector.
if NULL: test genome-wide genes
else: test the genes you input

2) when mode = "univariate", set `item` as NULL or a two-column matrix.
if NULL: test all ligand-receptor pair stored in our package
else: test the gene pairs you input
}
\examples{
genes <- c("TGFB1","TGFB2","TGFB3","TGFBR1","TGFBR2","TGFBR3")
genePairs <- data.frame(c("TGFB1","TGFB1"), c("TGFBR1","TGFBR2"))
SpaCET_obj <- SpaCET.SpatialCorrelation(SpaCET_obj, mode="univariate", item=genes, W, nPermutation=1000)
SpaCET_obj <- SpaCET.SpatialCorrelation(SpaCET_obj, mode="bivariate", item=genePairs, W, nPermutation=1000)
SpaCET_obj <- SpaCET.SpatialCorrelation(SpaCET_obj, mode="pairwise", W)

}
