% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{SpaCET.visualize.cellTypePair}
\alias{SpaCET.visualize.cellTypePair}
\title{Cell-type pair visualization}
\usage{
SpaCET.visualize.cellTypePair(SpaCET_obj, cellTypePair)
}
\arguments{
\item{SpaCET_obj}{An SpaCET object.}

\item{cellTypePair}{A pair of cell types.}
}
\value{
A ggplot object.
}
\description{
Visualize the interaction analysis of a co-localized cell-type pair.
}
\examples{
SpaCET.visualize.cellTypePair(SpaCET_obj, cellTypePair=c("CAF","Macrophage M2"))

}
