% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{SpaCET.visualize.colocalization}
\alias{SpaCET.visualize.colocalization}
\title{Cell-cell colocalization visualization}
\usage{
SpaCET.visualize.colocalization(SpaCET_obj)
}
\arguments{
\item{SpaCET_obj}{An SpaCET object.}
}
\value{
An ggplot object
}
\description{
Visualize the cell-type pair colocalization.
}
\examples{
SpaCET.visualize.colocalization(SpaCET_obj)
}
