% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SpaCET.visualize.deconvolution}
\alias{SpaCET.visualize.deconvolution}
\title{Cell type fraction visualization}
\usage{
SpaCET.visualize.deconvolution(
  SpaCET_obj,
  cellTypes,
  ncol = 1,
  colors = c("blue", "yellow", "red"),
  limits = c(0, 1),
  interactive = FALSE,
  imageBg = TRUE
)
}
\arguments{
\item{SpaCET_obj}{An SpaCET object.}

\item{colors}{Legend color scale, Default: c("blue", "yellow", "red").}

\item{limits}{Value range, Default: c(0,1). Also can be set as NULL.}

\item{cellType}{Cell type name.}
}
\value{
A ggplot2 object
}
\description{
Visualize cell type fraction in ST dataset.
}
\examples{
SpaCET.visualize.deconvolution(SpaCET_obj,"Malignant")
}
