% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SpaCET.visualize.gene}
\alias{SpaCET.visualize.gene}
\title{Gene expression visualization}
\usage{
SpaCET.visualize.gene(
  SpaCET_obj,
  genes,
  ncol = 1,
  colors = c("blue", "yellow", "red"),
  imageBg = TRUE
)
}
\arguments{
\item{SpaCET_obj}{An SpaCET object.}

\item{genes}{Gene symbol.}

\item{colors}{Legend color scale, Default: c("blue", "yellow", "red").}
}
\value{
A ggplot2 object
}
\description{
Visualize gene expression level in ST dataset.
}
\examples{
SpaCET.visualize.gene(SpaCET_obj,"EPCAM")
}
