% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SpaCET.visualize.LRNetworkScore}
\alias{SpaCET.visualize.LRNetworkScore}
\title{Ligand-Receptor network score visualization}
\usage{
SpaCET.visualize.LRNetworkScore(
  SpaCET_obj,
  colors = c("black", "black", "black", "blue", "blue", "blue", "blue", "cyan", "cyan",
    "yellow"),
  imageBg = TRUE
)
}
\arguments{
\item{SpaCET_obj}{An SpaCET object.}

\item{colors}{Legend color scale, Default: c("black","black","black","blue","blue","blue","blue","cyan","cyan","yellow").}
}
\value{
A ggplot2 object
}
\description{
Visualize L-R network score in ST dataset.
}
\examples{
SpaCET.visualize.LRNetworkScore(SpaCET_obj)
}
