% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SpaCET.visualize.metrics}
\alias{SpaCET.visualize.metrics}
\title{Quality control metrics visualization}
\usage{
SpaCET.visualize.metrics(
  SpaCET_obj,
  itemQC = c("UMI", "Gene"),
  colors = c("lightblue", "blue", "darkblue"),
  imageBg = TRUE,
  imageCut = c("complete", "cut_capture_area", "cut_tissue_region")
)
}
\arguments{
\item{SpaCET_obj}{An SpaCET object.}

\item{itemQC}{Item for quality control metrics. i.e., "UMI" or "gene".}

\item{colors}{Legend color scale, Default: c("blue", "yellow", "red").}

\item{imageBg}{logical: should the image be shown?}
}
\value{
A ggplot2 object
}
\description{
Visualize quality control metrics in ST dataset.
}
\examples{
SpaCET.visualize.metrics(SpaCET_obj, itemQC="UMI")
}
