% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.ST.R
\name{SpatialDeconv}
\alias{SpatialDeconv}
\title{ST Deconvolution step 2}
\usage{
SpatialDeconv(
  ST,
  Ref,
  malInput = TRUE,
  malProp,
  Unidentifiable = TRUE,
  MacrophageOther = TRUE,
  TCD4Other = TRUE
)
}
\arguments{
\item{ST}{PARAM_DESCRIPTION}

\item{Ref}{PARAM_DESCRIPTION}

\item{malInput}{PARAM_DESCRIPTION, Default: TRUE}

\item{malProp}{PARAM_DESCRIPTION}

\item{Unidentifiable}{PARAM_DESCRIPTION, Default: TRUE}

\item{MacrophageOther}{PARAM_DESCRIPTION, Default: TRUE}

\item{TCD4Other}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Hierarchical deconvolution of non-malignant cell fractions
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{constrOptim}}
}
