% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST.CCI.cellTypePair.R
\name{ST.CCI.cellTypePair.box}
\alias{ST.CCI.cellTypePair.box}
\title{Cell-type pair box plot}
\usage{
ST.CCI.cellTypePair.box(ST, cellTypePair)
}
\arguments{
\item{ST}{An SpaCE object}

\item{cellTypePair}{A vactor of two cell-types}
}
\value{
A ggplot2 object
}
\description{
Plot the L-R network score and p-value of a cell-type pair
}
\details{
This function plots the L-R network score and p-value of a cell-type pair (e.g., CAF and M2). SpaCE grouped ST spots into four categories, including CAF-M2 co-localization, CAF dominated, M2 dominated, and the rest.
}
\examples{
ST.CCI.cellTypePair.box(ST,cellTypePair=c("CAF","M2"))
}
