% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST.CCI.cellTypePair.R
\name{ST.CCI.cellTypePair.scatter}
\alias{ST.CCI.cellTypePair.scatter}
\title{FUNCTION_TITLE}
\usage{
ST.CCI.cellTypePair.scatter(ST, cellTypePair)
}
\arguments{
\item{ST}{PARAM_DESCRIPTION}

\item{cellTypePair}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[cowplot]{plot_grid}}
}
