% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST.CCI.colocalization.R
\name{ST.CCI.colocalization.plot}
\alias{ST.CCI.colocalization.plot}
\title{Cell colocalization plot}
\usage{
ST.CCI.colocalization.plot(ST)
}
\arguments{
\item{ST}{An SpaCE object}
}
\value{
A gggplot2 object
}
\description{
visualize the cell-cell colocalization
}
\details{
This function visualizes the cell colocalization in point and network plots.
}
\examples{

}
