% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST.CCI.LRInteraction.R
\name{ST.CCI.LRInteraction.plot}
\alias{ST.CCI.LRInteraction.plot}
\title{Ligand-Receptor interaction plot}
\usage{
ST.CCI.LRInteraction.plot(ST)
}
\arguments{
\item{ST}{An SpaCE object}
}
\value{
A ggplot2 object
}
\description{
Plot the Ligand-Receptor interaction with network score and p value
}
\details{
This function show the L-R network scores and p values for ST spots.
}
\examples{
ST.CCI.LRInteraction.plot(ST)
}
