% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST.CCI.LRInteraction.R
\name{ST.CCI.LRInteraction}
\alias{ST.CCI.LRInteraction}
\title{FUNCTION_TITLE}
\usage{
ST.CCI.LRInteraction(ST)
}
\arguments{
\item{ST}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[BiRewire]{birewire.rewire.bipartite}}
\code{\link[reshape2]{melt}}
}
