% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST.deconv.res.plot.R
\name{ST.deconv.res.scatter}
\alias{ST.deconv.res.scatter}
\title{FUNCTION_TITLE}
\usage{
ST.deconv.res.scatter(
  ST,
  cellTypes,
  cols = c("blue", "yellow", "red"),
  colsLimits = c(0, 1)
)
}
\arguments{
\item{ST}{PARAM_DESCRIPTION}

\item{cellTypes}{PARAM_DESCRIPTION}

\item{cols}{PARAM_DESCRIPTION, Default: c("blue", "yellow", "red")}

\item{colsLimits}{PARAM_DESCRIPTION, Default: c(0, 1)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[cowplot]{plot_grid}}
}
