% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST.deconv.res.plot.R
\name{ST.deconv.res.violin}
\alias{ST.deconv.res.violin}
\title{Deconvolution results violin plot}
\usage{
ST.deconv.res.violin(ST)
}
\arguments{
\item{ST}{An SpaCE object}
}
\value{
A ggplot2 object
}
\description{
Visualize the deconvolution results in violin plot
}
\details{
This function plot the deconvolution results across all ST spots
}
\examples{
ST.deconv.res.violin(ST)
}
