% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST.malignant.clone.R
\name{ST.malignant.clone.heatmap}
\alias{ST.malignant.clone.heatmap}
\title{Malignant cell clone heatmap}
\usage{
ST.malignant.clone.heatmap(ST)
}
\arguments{
\item{ST}{An SpaCE object}
}
\value{
An SpaCE object
}
\description{
Show the heatmap plot of inferred copy number variation values
}
\details{
This function obtains the heatmap plot of inferred copy number variation values.
}
\examples{

}
