% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST.metrics.R
\name{ST.metrics.plot}
\alias{ST.metrics.plot}
\title{FUNCTION_TITLE}
\usage{
ST.metrics.plot(ST, cols = c("lightblue", "blue", "darkblue"))
}
\arguments{
\item{ST}{PARAM_DESCRIPTION}

\item{cols}{PARAM_DESCRIPTION, Default: c("lightblue", "blue", "darkblue")}

\item{colsLimits}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[cowplot]{plot_grid}}
}
