% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST.metrics.R
\name{ST.metrics}
\alias{ST.metrics}
\title{FUNCTION_TITLE}
\usage{
ST.metrics(ST)
}
\arguments{
\item{ST}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
