% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{visualSpatial}
\alias{visualSpatial}
\title{FUNCTION_TITLE}
\usage{
visualSpatial(
  visiualVector,
  HEimage,
  cols,
  colsLimits,
  titleName,
  legendName,
  legendPosition = "right",
  is.continuous = TRUE
)
}
\arguments{
\item{visiualVector}{PARAM_DESCRIPTION}

\item{HEimage}{PARAM_DESCRIPTION}

\item{cols}{PARAM_DESCRIPTION}

\item{colsLimits}{PARAM_DESCRIPTION}

\item{titleName}{PARAM_DESCRIPTION}

\item{legendName}{PARAM_DESCRIPTION}

\item{is.continuous}{PARAM_DESCRIPTION}

\item{lgendPosition}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[png]{readPNG}}
\code{\link[grid]{grid.raster}},\code{\link[grid]{unit}}
}
