% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{write.gmt}
\alias{write.gmt}
\title{Write a gmt file.}
\usage{
write.gmt(gmt, gmtPath)
}
\arguments{
\item{gmt}{A gene set list.}

\item{gmtPath}{Path to a gmt file.}
}
\description{
Write a gene set list as a gmt file.
}
\examples{
write.gmt(gmt, gmtPath)

}
