% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{addTo.Seurat}
\alias{addTo.Seurat}
\title{Add SpaCET to Seurat}
\usage{
addTo.Seurat(SpaCET_obj, Seurat_obj)
}
\arguments{
\item{SpaCET_obj}{An SpaCET object.}

\item{Seurat_obj}{An Seurat object.}
}
\value{
An Seurat object.
}
\description{
Add deconvolution results from an SpaCET object to an Seurat object as a new assay.
}
\examples{
visiumPath <- file.path(system.file(package = "SpaCET"), "extdata/Visium_BC")
Seurat_obj <- Load10X_Spatial(data.dir = visiumPath)
SpaCET_obj <- convert.Seurat(Seurat_obj)
SpaCET_obj <- SpaCET.deconvolution(SpaCET_obj, cancerType="BRCA", coreNo=8)
Seurat_obj <- addTo.Seurat(SpaCET_obj, Seurat_obj)
DefaultAssay(Seurat_obj) <- "propMatFromSpaCET"
SpatialFeaturePlot(Seurat_obj, features = c("CAF", "Macrophage"))
}
