% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert.Seurat}
\alias{convert.Seurat}
\title{Convert Seurat to SpaCET}
\usage{
convert.Seurat(Seurat_obj)
}
\arguments{
\item{Seurat_obj}{An Seurat object.}
}
\value{
An SpaCET object.
}
\description{
Convert an Seurat object to an SpaCET object.
}
\examples{
visiumPath <- file.path(system.file(package = "SpaCET"), "extdata/Visium_BC")
Seurat_obj <- Seurat::Load10X_Spatial(data.dir = visiumPath)
SpaCET_obj <- convert.Seurat(Seurat_obj)
SpaCET_obj <- SpaCET.quality.control(SpaCET_obj)
SpaCET.visualize.spatialFeature(SpaCET_obj, spatialType = "QualityControl", spatialFeatures=c("UMI","Gene"))

}
