% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.SpaCE.R
\name{create.SpaCE}
\alias{create.SpaCE}
\title{Data Input}
\usage{
create.SpaCE(seqPath, imagePath)
}
\arguments{
\item{seqPath}{Path to the folder 'sequencing'.}

\item{imagePath}{Path to the folder 'image'.}
}
\value{
An SpaCE object
}
\description{
Read ST data set to create an SpaCE object.
}
\details{
Basically, \code{create.SpaCE} requires two parameters \code{seqPath} and \code{imagePath}, which point the standard output folders of 10x Visium data.

The \code{seqPath} folder should include \cr
"barcodes.tsv.gz": spot level barcodes; \cr
"features.tsv.gz": list of genes; \cr
"matrix.mtx.gz": (sparse) matrix of counts.

The \code{imagePath} folder should include \cr
“tissue_positions_list.csv” : barcodes and spatial information; \cr
“tissue_lowres_image.png” : hematoxylin and eosin (H&E) image; \cr
“scalefactors_json.json” : scaling factors for adjusting the coordinates .
}
\examples{
ST <- create.SpaCE(seqPath,imagePath)
}
