% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create.SpaCET.object.10X}
\alias{create.SpaCET.object.10X}
\title{Create an SpaCET object from 10X Visium}
\usage{
create.SpaCET.object.10X(visiumPath)
}
\arguments{
\item{visiumPath}{Path to the Space Ranger output folder. See ‘details’ for more information.}
}
\value{
An SpaCET object.
}
\description{
Read an ST dataset to create an SpaCET object.
}
\details{
If users are analyzing an ST data set from 10X Visium platform, they only need to input "visiumPath".
Please make sure that "visiumPath" points to the standard output folder of 10X Space Ranger,
which have both `filtered_feature_bc_matrix` and `spatial` folders.

The "filtered_feature_bc_matrix" folder includes \cr
"barcodes.tsv.gz": spot level barcodes; \cr
"features.tsv.gz": list of genes; \cr
"matrix.mtx.gz": (sparse) matrix of counts.

The "spatial" folder includes \cr
“tissue_positions_list.csv” : barcodes and spatial information; \cr
“tissue_lowres_image.png” : hematoxylin and eosin (H&E) image; \cr
“scalefactors_json.json” : scaling factors for adjusting the coordinates.
}
\examples{
visiumPath <- file.path(system.file(package = "SpaCET"), "extdata/Visium_BC")
SpaCET_obj <- create.SpaCET.object.10X(visiumPath = visiumPath)

}
