% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.ST.R
\name{deconvolve.ST}
\alias{deconvolve.ST}
\title{ST Deconvolution}
\usage{
deconvolve.ST(ST, tempFilePath)
}
\arguments{
\item{ST}{An SpaCE object}

\item{tempFilePath}{Path to a temporary folder for storing infercnv results.}
}
\value{
An SpaCE object
}
\description{
Deconvovle tumor ST data hierarchically
}
\details{
This function first estimates cancer cell abundance through modeling the segmental copy number variations across ST spots. Then, a constrained regression model calibrates local tissue densities and determines stromal and immune cell lineage hierarchies.
}
\examples{

}
