% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.ST.R
\name{deconvolve.ST}
\alias{deconvolve.ST}
\title{FUNCTION_TITLE}
\usage{
deconvolve.ST(ST, tempFilePath)
}
\arguments{
\item{ST}{PARAM_DESCRIPTION}

\item{tempFilePath}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
