% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaCE-package.R
\docType{package}
\name{SpaCE-package}
\alias{SpaCE}
\alias{SpaCE-package}
\title{SpaCE: Spatial Cellular Estimator}
\description{
Estimation of cell lineage and interactions in tumors from spatial transcriptomics data.
}
\author{
\strong{Maintainer}: Beibei Ru \email{first.last@example.com}

}
\keyword{internal}
