% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{SpaCE.deconvolution}
\alias{SpaCE.deconvolution}
\title{Deconvolve tumor ST data set}
\usage{
SpaCE.deconvolution(SpaCE_obj, cancerType, coreNo = 8)
}
\arguments{
\item{SpaCE_obj}{An SpaCE object.}

\item{cancerType}{Cancer type of this tumor ST dataset.}

\item{coreNo}{Core number.}
}
\value{
An SpaCE object
}
\description{
Estimate the fraction of cell lineage and sub lineage.
}
\examples{
SpaCE_obj <- SpaCE.deconvolution(SpaCE_obj, cancerType="BRCA", coreNo=8)
}
