% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SpaCE.visualize.deconvolution}
\alias{SpaCE.visualize.deconvolution}
\title{Cell type fraction visualization}
\usage{
SpaCE.visualize.deconvolution(
  SpaCE_obj,
  cellType,
  colors = c("blue", "yellow", "red"),
  limits = c(0, 1),
  interactive = FALSE,
  imageBg = TRUE
)
}
\arguments{
\item{SpaCE_obj}{An SpaCE object.}

\item{cellType}{Cell type name.}

\item{colors}{Legend color scale, Default: c("blue", "yellow", "red").}

\item{limits}{Value range, Default: c(0,1). Also can be set as NULL.}
}
\value{
A ggplot2 object
}
\description{
Visualize cell type fraction in ST dataset.
}
\examples{
SpaCE.visualize.deconvolution(SpaCE_obj,"Malignant")
}
