% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SpaCE.visualize.gene}
\alias{SpaCE.visualize.gene}
\title{Gene expression visualization}
\usage{
SpaCE.visualize.gene(
  SpaCE_obj,
  gene,
  colors = c("blue", "yellow", "red"),
  imageBg = TRUE
)
}
\arguments{
\item{SpaCE_obj}{An SpaCE object.}

\item{colors}{Legend color scale, Default: c("blue", "yellow", "red").}

\item{geneSymbol}{Gene symbol.}
}
\value{
A ggplot2 object
}
\description{
Visualize gene expression level in ST dataset.
}
\examples{
SpaCE.visualize.gene(SpaCE_obj,"EPCAM")
}
