% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SpaCE.visualize.LRNetworkScore}
\alias{SpaCE.visualize.LRNetworkScore}
\title{Ligand-Receptor network score visualization}
\usage{
SpaCE.visualize.LRNetworkScore(
  SpaCE_obj,
  colors = c("black", "#081d58", "#253494", "blue", "blue", "blue", "blue", "cyan",
    "cyan", "yellow"),
  interactive = FALSE,
  imageBg = TRUE
)
}
\arguments{
\item{SpaCE_obj}{An SpaCE object.}

\item{colors}{Legend color scale, Default: c("blue", "yellow", "red").}

\item{limits}{Value range, Default: c(0,2).}
}
\value{
A ggplot2 object
}
\description{
Visualize L-R network score in ST dataset.
}
\examples{
SpaCE.visualize.LRNetworkScore(SpaCE_obj)
}
