% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{SpaCET.deconvolution.malignant.customized.scRNAseq}
\alias{SpaCET.deconvolution.malignant.customized.scRNAseq}
\title{Deconvolve malignant fraction with customized scRNAseq data}
\usage{
SpaCET.deconvolution.malignant.customized.scRNAseq(
  SpaCET_obj,
  Malignant = "Malignant",
  sc_counts,
  sc_annotation,
  sc_lineageTree,
  sc_nCellEachLineage = 100,
  coreNo = 6
)
}
\arguments{
\item{SpaCET_obj}{A SpaCET object.}

\item{Malignant}{Indicates the name of malignant cell type in the major lineage layer from the deconvolution results. Default: "Malignant".}

\item{sc_counts}{Single cell count matrix with gene name (row) x cell ID (column).}

\item{sc_annotation}{Single cell annotation matrix. This matrix should include two columns, i,e., cellID and cellType. Each row represents a single cell.}

\item{sc_lineageTree}{Cell lineage tree. This should be organized by using a list, and the name of each element are major lineages while the value of elements are the corresponding sublineages. If a major lineage does not have any sublineages, the value of this major lineage should be itself.}

\item{sc_nCellEachLineage}{Cell count each lineage. Default: 100. If a cell type is comprised of >100 cells, only 100 cells per cell identity are randomly selected to generate cell type reference.}

\item{coreNo}{Core number in parallel.}
}
\value{
A SpaCET object
}
\description{
Explore different malignant cell states in tumor ST dataset.
}
\examples{
SpaCET_obj <- SpaCET.deconvolution.malignant.customized.scRNAseq(SpaCET_obj, Malignant="Malignant", sc_counts, sc_annotation, sc_lineageTree, sc_nCellEachLineage=100, coreNo=6)

}
