% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{SpaCET.deconvolution.malignant}
\alias{SpaCET.deconvolution.malignant}
\title{Deconvolve malignant cell fraction}
\usage{
SpaCET.deconvolution.malignant(
  SpaCET_obj,
  Malignant = "Malignant",
  malignantCutoff = 0.7,
  coreNo = 6
)
}
\arguments{
\item{SpaCET_obj}{A SpaCET object.}

\item{Malignant}{Indicates the name of malignant cell type in the major lineage layer from the deconvolution results. Default: "Malignant".}

\item{malignantCutoff}{Fraction cutoff for defining spots with high abundant malignant cells. Default: 0.7.}

\item{coreNo}{Core number in parallel.}
}
\value{
A SpaCET object
}
\description{
Explore different malignant cell states in tumor ST dataset.
}
\examples{
SpaCET_obj <- SpaCET.deconvolution.malignant(SpaCET_obj)

}
