% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{SpaCET.deconvolution.malignant}
\alias{SpaCET.deconvolution.malignant}
\title{Deconvolve malignant cell fraction}
\usage{
SpaCET.deconvolution.malignant(SpaCET_obj, malignantCutoff = 0.7, coreNo = 8)
}
\arguments{
\item{SpaCET_obj}{An SpaCET object.}

\item{malignantCutoff}{Fraction cutoff for defining spots with high abundant malignant cells.}

\item{coreNo}{Core number in parallel.}
}
\value{
An SpaCET object
}
\description{
Explore different cancer cell state in tumor ST dataset.
}
\examples{
SpaCET_obj <- SpaCET.deconvolution.malignant(SpaCET_obj)

}
