% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{SpaCET.deconvolution.matched.scRNAseq}
\alias{SpaCET.deconvolution.matched.scRNAseq}
\title{Deconvolve ST data set with matched scRNAseq}
\usage{
SpaCET.deconvolution.matched.scRNAseq(
  SpaCET_obj,
  sc_counts,
  sc_annotation,
  sc_lineageTree,
  coreNo = 8
)
}
\arguments{
\item{SpaCET_obj}{An SpaCET object.}

\item{sc_counts}{Single cell count matrix with gene name (row) x cell ID (column).}

\item{sc_annotation}{Single cell annotation matrix. This matrix should include two columns, i,e., cellID and cellType. Each row represents a single cell.}

\item{sc_lineageTree}{Cell lineage tree. This should be organized by using a list, and the name of each element are major lineages while the value of elements are the corresponding sublineages. If a major lineage does not have any sublineages, the value of this major lineage should be itself.}

\item{coreNo}{Core number.}
}
\value{
An SpaCET object
}
\description{
Estimate the fraction of cell lineage and sub lineage.
}
\examples{
SpaCET_obj <- SpaCET.deconvolution.matched.scRNAseq(SpaCET_obj, sc_counts, sc_annotation, sc_lineageTree)

}
