% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{SpaCET.deconvolution.matched.scRNAseq}
\alias{SpaCET.deconvolution.matched.scRNAseq}
\title{Deconvolve ST data set with matched scRNAseq data}
\usage{
SpaCET.deconvolution.matched.scRNAseq(
  SpaCET_obj,
  sc_includeMalignant = TRUE,
  cancerType,
  sc_counts,
  sc_annotation,
  sc_lineageTree,
  sc_downsampling = TRUE,
  sc_nCellEachLineage = 100,
  coreNo = 6
)
}
\arguments{
\item{SpaCET_obj}{A SpaCET object.}

\item{sc_includeMalignant}{Indicate whether the single cell data include malignant cells. If FALSE, please input a cancer type and then SpaCET will infer the malignant cell fraction based on its build-in reference.}

\item{cancerType}{Cancer type of the current tumor ST sample.}

\item{sc_counts}{Single cell count matrix with gene name (row) x cell ID (column).}

\item{sc_annotation}{Single cell annotation matrix. This matrix should include two columns, i,e., cellID and cellType. Each row represents a single cell.}

\item{sc_lineageTree}{Cell lineage tree. This should be organized by using a list, and the name of each element are major lineages while the value of elements are the corresponding sublineages. If a major lineage does not have any sublineages, the value of this major lineage should be itself.}

\item{sc_downsampling}{Indicate whether downsample the single cell data for each cell type. Default: TRUE.}

\item{sc_nCellEachLineage}{Cell count each lineage. Default: 100. If a cell type is comprised of >100 cells, only 100 cells per cell identity are randomly selected to generate cell type reference.}

\item{coreNo}{Core number.}
}
\value{
A SpaCET object
}
\description{
Estimate the fraction of cell lineage and sub lineage.
}
\examples{
SpaCET_obj <- SpaCET.deconvolution.matched.scRNAseq(SpaCET_obj, sc_counts, sc_annotation, sc_lineageTree)

}
