% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{SpaCET.deconvolution}
\alias{SpaCET.deconvolution}
\title{Deconvolve tumor ST data set}
\usage{
SpaCET.deconvolution(SpaCET_obj, cancerType, coreNo = 8)
}
\arguments{
\item{SpaCET_obj}{An SpaCET object.}

\item{cancerType}{Cancer type of this tumor ST dataset.}

\item{coreNo}{Core number in parallel.}
}
\value{
An SpaCET object
}
\description{
Estimate the fraction of cell lineages and sub lineages.
}
\examples{
SpaCET_obj <- SpaCET.deconvolution(SpaCET_obj, cancerType="BRCA", coreNo=8)
}
