% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{SpaCET.deconvolution}
\alias{SpaCET.deconvolution}
\title{Deconvolve tumor ST data set}
\usage{
SpaCET.deconvolution(
  SpaCET_obj,
  cancerType,
  adjacentNormal = FALSE,
  coreNo = 6
)
}
\arguments{
\item{SpaCET_obj}{A SpaCET object.}

\item{cancerType}{Cancer type of the current tumor ST dataset.}

\item{adjacentNormal}{Indicate whether your sample is normal tissue adjacent to the tumor. If TURE, SpaCET will skip the stage of malignant cell inference. Default: FALSE.}

\item{coreNo}{Core number in parallel computation.}
}
\value{
A SpaCET object.
}
\description{
Estimate the cell fraction of cell lineages and sub lineages.
}
\examples{
SpaCET_obj <- SpaCET.deconvolution(SpaCET_obj, cancerType="BRCA", coreNo=6)

}
