% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{SpaCET.identify.interface}
\alias{SpaCET.identify.interface}
\title{Identify tumor-stroma interface}
\usage{
SpaCET.identify.interface(
  SpaCET_obj,
  Malignant = "Malignant",
  MalignantCutoff = 0.5
)
}
\arguments{
\item{SpaCET_obj}{An SpaCET object.}

\item{Malignant}{Indicates the name of malignant cell type in the major lineage layer from the deconvolution results. Default: "Malignant".}

\item{MalignantCutoff}{Malignant cell fraction cutoff for tumor boundary. Default: 0.5.}
}
\value{
An SpaCET object.
}
\description{
Identify the spots at the tumor-stroma interface.
}
\examples{
SpaCET_obj <- SpaCET.identify.interface(SpaCET_obj)
SpaCET.visualize.spatialFeature(SpaCET_obj, spatialType = "Interface")

}
