% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{SpaCET.quality.control}
\alias{SpaCET.quality.control}
\title{Filter spatial spots and calculate the QC metrics}
\usage{
SpaCET.quality.control(SpaCET_obj, min.genes = 1)
}
\arguments{
\item{min.genes}{Minimum number of expressed genes. Default: 1.}

\item{SpaCET}{An SpaCET object.}
}
\value{
An SpaCET object.
}
\description{
Spots with less than `min.genes` expressed genes would be removed.
}
\examples{
SpaCET_obj <- SpaCET.quality.control(SpaCET_obj)
}
