% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SpaCET.visualize.spatialFeature}
\alias{SpaCET.visualize.spatialFeature}
\title{Spatial feature visualization}
\usage{
SpaCET.visualize.spatialFeature(
  SpaCET_obj,
  spatialType = c("QualityControl", "GeneExpr", "CellFraction", "LRNetworkScore"),
  spatialFeatures = NULL,
  pointSize = 1,
  nrow = 1,
  imageBg = TRUE
)
}
\arguments{
\item{SpaCET_obj}{An SpaCET object.}

\item{spatialType}{Type of spatial features, i.e., "QualityControl", "GeneExpression", "CellFraction", and "LRNetworkScore". See ‘details’ for more information.}

\item{spatialFeatures}{A vector of spatial features.}

\item{nrow}{Row number of the combined panel for multiple spatial features.}

\item{imageBg}{logical: should the image be shown?}
}
\value{
A ggplot2 object.
}
\description{
Visualize multiple types of spatial features in ST data.
}
\details{
`SpaCET.visualize.spatialFeature` is able to plot multiple types of spatial features, including "QC", "GeneExp", "CellFraction", and "LRNetworkScore".
"QualityControl" refers to quality control metrics. User can visualize the total counts of UMI and genes across all spots.
"GeneExpr" is set to visualize the expression level of genes.
"CellFraction" refers to the cell fraction of cell types.
"LRNetworkScore" is selected to show the Ligand-Receptor network score. See `SpaCET.CCI.LRNetworkScore` for how to calculate it.
}
\examples{
SpaCET.visualize.spatialFeature(SpaCET_obj, spatialType = "QualityControl", spatialFeatures=c("UMI","Gene"))
SpaCET.visualize.spatialFeature(SpaCET_obj, spatialType = "GeneExpression", spatialFeatures=c("EPCAM","MS4A1"))
SpaCET.visualize.spatialFeature(SpaCET_obj, spatialType = "CellFraction", spatialFeatures=c("Malignant","CAF"))
SpaCET.visualize.spatialFeature(SpaCET_obj, spatialType = "CellFraction", spatialFeatures="All", pointSize = 0.1, nrow=5)
SpaCET.visualize.spatialFeature(SpaCET_obj, spatialType = "LRNetworkScore", spatialFeatures=c("Network_Score","Network_Score_pv"))

}
