% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SpaCET.visualize.spatialFeature}
\alias{SpaCET.visualize.spatialFeature}
\title{Spatial feature visualization}
\usage{
SpaCET.visualize.spatialFeature(
  SpaCET_obj,
  spatialType,
  spatialFeatures = NULL,
  scaleTypeForGeneExpression = "LogTPM",
  sameScaleForFraction = FALSE,
  colors = NULL,
  pointSize = 1,
  pointAlpha = 1,
  nrow = 1,
  imageBg = TRUE,
  imageSize = "CaptureArea",
  CustomizedAreaScale = NULL,
  legend.position = "right",
  interactive = FALSE
)
}
\arguments{
\item{SpaCET_obj}{A SpaCET object.}

\item{spatialType}{Type of spatial features, i.e., "QualityControl", "GeneExpression", "CellFraction", and "LRNetworkScore". See ‘details’ for more information.}

\item{spatialFeatures}{A vector of spatial features.}

\item{scaleTypeForGeneExpression}{Scale type of gene expression, i.e., "RawCounts","LogRawCounts","LogTPM/10", and "LogTPM".}

\item{sameScaleForFraction}{Indicate whether all cell types have the same scale for cell fraction.}

\item{colors}{A vector of colors for spots.}

\item{pointSize}{Size of spots.}

\item{pointAlpha}{Alpha transparency scales of spots. Must lie between 0 and 1.}

\item{nrow}{Row number of the combined panel for multiple spatial features.}

\item{imageBg}{Logical: should the image be shown as background?}

\item{imageSize}{Size of the image, i.e., "CompleteImage", "CaptureArea", "CustomizedArea".}

\item{CustomizedAreaScale}{A vector of four numbers (0~1) for scale of the Customized Area, i.e., x_left, x_right, y_bottom, y_top.}

\item{legend.position}{The position of the legend. Set it as "none" if you want to remove the legend.}

\item{interactive}{Logical: should the interactive app be activited?}
}
\value{
A ggplot2 object.
}
\description{
Visualize multiple types of spatial features in ST data.
}
\details{
`SpaCET.visualize.spatialFeature` is able to plot multiple types of spatial features, including "QC", "GeneExp", "CellFraction", and "LRNetworkScore".
"QualityControl" refers to quality control metrics. User can visualize the total counts of UMI and genes across all spots.
"GeneExpression" is set to visualize the expression level of genes.
"CellFraction" refers to the cell fraction of cell types.
"LRNetworkScore" is selected to show the Ligand-Receptor network score. See `SpaCET.CCI.LRNetworkScore` for how to calculate it.
}
\examples{
SpaCET.visualize.spatialFeature(SpaCET_obj, spatialType = "QualityControl", spatialFeatures=c("UMI","Gene"))
SpaCET.visualize.spatialFeature(SpaCET_obj, spatialType = "GeneExpression", spatialFeatures=c("EPCAM","MS4A1"))
SpaCET.visualize.spatialFeature(SpaCET_obj, spatialType = "CellFraction", spatialFeatures=c("Malignant","Macrophage"))
SpaCET.visualize.spatialFeature(SpaCET_obj, spatialType = "LRNetworkScore", spatialFeatures=c("Network_Score","Network_Score_pv"))

}
