% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST.CCI.colocalization.R
\name{ST.CCI.colocalization}
\alias{ST.CCI.colocalization}
\title{Cell colocalization}
\usage{
ST.CCI.colocalization(ST)
}
\arguments{
\item{ST}{An SpaCE object}
}
\value{
An SpaCE object
}
\description{
Calculate the cell-cell colocalization
}
\details{
This function calculates the Spearman correlation of cell types as colocalization.
}
\examples{

}
