% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST.CCI.LRInteraction.R
\name{ST.CCI.LRInteraction}
\alias{ST.CCI.LRInteraction}
\title{Ligand-Receptor interaction}
\usage{
ST.CCI.LRInteraction(ST)
}
\arguments{
\item{ST}{An SpaCE object}
}
\value{
An SpaCE object
}
\description{
Calculate the Ligand-Receptor interaction with network score and p value
}
\details{
SpaCE computes an L-R interactions network score for each spot as the sum of expression products between L-R pairs, normalized by the average L-R network score from 1,000 random networks with the same degrees. The p-value is calculated as the fraction of random network scores that exceeded the original score.
}
\examples{
ST <- ST.CCI.LRInteraction(ST)
}
