% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST.deconv.res.plot.R
\name{ST.deconv.res.scatter}
\alias{ST.deconv.res.scatter}
\title{Deconvolution results scatter plot}
\usage{
ST.deconv.res.scatter(
  ST,
  cellTypes,
  cols = c("blue", "yellow", "red"),
  colsLimits = c(0, 1)
)
}
\arguments{
\item{ST}{An SpaCE object}

\item{cellTypes}{A vector of cell-types}

\item{cols}{Colors, Default: c("blue", "yellow", "red")}

\item{colsLimits}{Colors Limits of cell type fraction, Default: c(0, 1)}
}
\value{
A ggplot2 object
}
\description{
Visualize the deconvolution results in scatter plot
}
\details{
DETAILS
}
\examples{
ST.deconv.res.scatter(ST, cellTypes=c("Malignant","B cell","CAF") )
}
