% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST.deconv.res.plot.R
\name{ST.deconv.res.violin}
\alias{ST.deconv.res.violin}
\title{FUNCTION_TITLE}
\usage{
ST.deconv.res.violin(ST)
}
\arguments{
\item{ST}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[reshape]{melt}}
}
