% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST.malignant.clone.R
\name{ST.malignant.clone.heatmap}
\alias{ST.malignant.clone.heatmap}
\title{Malignant cell clone heatmap}
\usage{
ST.malignant.clone.heatmap(ST, cutoffMalignant, nClone)
}
\arguments{
\item{ST}{An SpaCE object}

\item{cutoffMalignant}{Number ranging 0-1}

\item{nClone}{Integer}
}
\value{
An SpaCE object
}
\description{
Cluster the inferred copy number variations to indentify malignant cell clone
}
\details{
This function carries out hierarchical clustering of inferred copy number variation values.
}
\examples{
ST <- ST.malignant.clone.heatmap(ST, cutoffMalignant=0.5, nClone=3)
}
