% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST.malignant.clone.R
\name{ST.malignant.clone.heatmap}
\alias{ST.malignant.clone.heatmap}
\title{FUNCTION_TITLE}
\usage{
ST.malignant.clone.heatmap(ST, cutoffMalignant, nClone)
}
\arguments{
\item{ST}{PARAM_DESCRIPTION}

\item{cutoffMalignant}{PARAM_DESCRIPTION}

\item{nClone}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
