% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST.malignant.clone.R
\name{ST.malignant.clone.scatter}
\alias{ST.malignant.clone.scatter}
\title{Malignant cell clone scatter plot}
\usage{
ST.malignant.clone.scatter(ST, cols)
}
\arguments{
\item{ST}{An SpaCE object}

\item{cols}{Specify a custom set of colors for clones}
}
\value{
A ggplot2 object
}
\description{
Show the spatial distribution of malignant cell clones
}
\details{
This function obtains the spatial distribution of distinct malignant cell clones.
}
\examples{

}
