% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST.metrics.R
\name{ST.metrics.plot}
\alias{ST.metrics.plot}
\title{ST metrics plot}
\usage{
ST.metrics.plot(ST, cols = c("lightblue", "blue", "darkblue"))
}
\arguments{
\item{ST}{An SpaCE object}

\item{cols}{Colors, Default: c("lightblue", "blue", "darkblue")}
}
\value{
A ggplot2 object
}
\description{
Plot the quality control metrics for all ST spots
}
\examples{
ST.metrics.plot(ST, cols = c("lightblue", "blue", "darkblue"))
}
