% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST.metrics.R
\name{ST.metrics}
\alias{ST.metrics}
\title{ST metrics}
\usage{
ST.metrics(ST)
}
\arguments{
\item{ST}{An SpaCE object}
}
\value{
An SpaCE object
}
\description{
Calculate the quality control metrics
}
\details{
This function computes both UMI and gene counts across all ST spots
}
\examples{
ST <- ST.metrics(ST)
}
